p2p: sentinel the per-IP descriptor, and stop restating the plan doc
Four review findings, plus a stale pin.

F-1 (medium, latent). The descriptor at net_node.cpp:191 initialized an
extern const (net_node.h:793) with a function call, which forces DYNAMIC
initialization where the literal gave constant initialization. Its
zero-initialized value is 0, and cap 0 admits nothing -- so any future
static-init reader gets a silent, total inbound outage. Nothing reads it
during static init today, so this was latent rather than live, but the
precedent the PR cites does not do it: shekyl_p2p_default_out_peers() is
called only from function bodies.

Fixed the way --out-peers and --in-peers already do it: the descriptor
carries the SENTINEL -1, and handle_command_line resolves it. The type is
now int64_t so it CAN carry a sentinel -- 0 is a legal operator choice
meaning "refuse every inbound connection", so it cannot double as unset.
That also removes the last C++ literal, so the ownership claim gets
stronger rather than merely patched: the default now exists in exactly one
place. A test pins the trap directly -- sentinel resolves to the Rust
default, and 0 survives as 0 -- with the red edit named (`<= 0` in the
resolution).

F-1's second half: the constraint was unwritten. shekyl_host_inbound_default_cap
must stay a constant return with no Rust-side global or lazy state, because
the node_server constructor calls it and a OnceLock implementation would be
a genuine initialization-order fault whose failure mode is silent. Stated on
the Rust entry point and mirrored in shekyl_ffi.h, since the C++ half cannot
enforce it.

F-2. The commit claimed a ruling moving the number has "one place to land",
while node_server.cpp and the Rust test both hard-pinned 1 -- three files.
The Rust pin earns its place as the did-not-quietly-change check; the C++
literal duplicated it, and that test's real job is the agreement assertion
beside it, which is value-independent. Literal dropped; the claim is now
true.

F-3. FOLLOWUPS.md's header says essays live in the owning plan doc. Until
today these rows had no owner to point at; they do now, and PWD-I7/PWD-I8
arrived with the full argument. Two copies of one argument is two things to
keep current. Both rows trimmed to claim, anchors and falsifier, with the
narrative handed to I7/I8.

F-4 was retracted by the reviewer and needs no change: the record shows one
node refused, not two -- A synced at 7101 holding the slots, B refused
across 34 handshakes -- which is what cap 1 predicts.

Stale pin. Anchors were pinned to f6df3abc2; dev has moved to 059aca264
(#808, #810). All eleven re-verified line by line at the new SHA and every
one still resolves, so this is a pin refresh and not a correction.

Falsifier scope, stated because the cheap outcome is the dangerous one. The
run RE-SEQUENCES PWD-I8 and cannot close it. The cap is already shown to be
what refuses B; what the falsifier settles is whether the failure cache is
what makes the refusal unrecoverable. I8 exists because is_same_host answers
two questions with one quantity, which holds whichever mechanism produced
this dead node -- so a green falsifier leaves a fixed incident and an
unfixed category error, and a working node is the thing most likely to stop
anyone looking again. Methodology recorded: raise every live seed, restart
BOTH daemons to clear m_conn_fails_cache (B's restart matters more than A's,
since B carries 34 recorded failures), and read the observation off the seed
rather than off either daemon's sync progress.

+++ src/p2p/net_node.h
@@ -790,7 +790,7 @@ namespace nodetool
     extern const command_line::arg_descriptor<int64_t> arg_limit_rate_down;
     extern const command_line::arg_descriptor<int64_t> arg_limit_rate;
     extern const command_line::arg_descriptor<bool> arg_pad_transactions;
-    extern const command_line::arg_descriptor<uint32_t> arg_max_connections_per_ip;
+    extern const command_line::arg_descriptor<int64_t> arg_max_connections_per_ip;
 }
 
 POP_WARNINGS
